change line filtering in forms
due to machine text being inconsistent with its use of hyphons at the end of lines, I only filter out lines in the beginning and the end of the form that don't occur in the machine part. This is to avoid handwritten text that is not part of machine text - see form p02-109 for an example.